Extracted components from PosePreset
