+++ src/js/shot-generator-new/utils/filepathFor.js
@@ -1,0 +1,7 @@
+
+import ModelLoader from '../../services/model-loader'
+const filepathFor = model => 
+  ModelLoader.getFilepathForModel(
+    { model: model.id, type: model.type },
+    { storyboarderFilePath: null })
+export { filepathFor }
