+++ src/js/shot-generator-new/utils/searchPresetsForTerms.js
@@ -1,0 +1,33 @@
+import LiquidMetal from 'liquidmetal'
+
+const comparePresetNames = (a, b) => {
+  var nameA = a.name.toUpperCase()
+  var nameB = b.name.toUpperCase()
+
+  if (nameA < nameB) {
+    return -1
+  }
+  if (nameA > nameB) {
+    return 1
+  }
+  return 0
+}
+const comparePresetPriority = (a, b) => b.priority - a.priority
+
+const searchPresetsForTerms = (presets, terms) => {
+  const matchAll = terms == null || terms.length === 0
+
+  return presets
+    .sort(comparePresetNames)
+    .filter(preset => {
+      if (matchAll) return true
+
+      return (
+        (LiquidMetal.score(preset.name, terms) > 0.8) ||
+        (preset.keywords && LiquidMetal.score(preset.keywords, terms) > 0.8)
+      )
+    })
+    .sort(comparePresetPriority)
+}
+
+export {searchPresetsForTerms}
